apply_repulsion_3d_simd_f32_parallel

+++ examples/viz/src/main.rs
@@ -7,7 +7,7 @@ use std::{io::BufRead, sync::Arc, thread, time::Duration};
 
 const STANDBY_SLEEP: Duration = Duration::from_millis(50);
 const COMPUTE_SLEEP: Duration = Duration::from_millis(1);
-type T = f64;
+type T = f32;
 
 fn main() {
 	let file = std::fs::File::open(std::env::args().nth(1).expect("Usage: viz <csv_file>"))
